node-fetch for server.js
