+++ package.json
@@ -16,6 +16,7 @@
     "jsonwebtoken": "^8.5.1",
     "moment": "^2.24.0",
     "morgan": "^1.9.1",
+    "node-fetch": "^2.6.0",
     "nodemon": "^1.19.4",
     "pg": "^7.12.1",
     "string-hash": "^1.1.3",
